Index file created
The index.html file was created with boilerplate code and a heading Odin recipes.

Recipe folder was created

The recipe folder was created and under this folder three html files were created containing the name of each recipe as heading.

Recipes are linked to index.html

All the recipes were linked to index.html file in three anchor tags followed by the name of recipes.